+++ apps/api/src/app/user/user.service.ts
@@ -220,7 +220,6 @@ export class UserService {
         return {
           alias: accessItem.alias,
           id: accessItem.id,
-          permissions: accessItem.permissions,
           scopes: getScopesOfAccess(accessItem)
         };
       }),