+++ libs/common/src/lib/dtos/create-access.dto.ts
@@ -1,7 +1,7 @@
 import { Filter } from '@ghostfolio/common/interfaces';
+import { Scope, scopes } from '@ghostfolio/common/scopes';
 
-import { AccessPermission } from '@prisma/client';
-import { IsArray, IsEnum, IsOptional, IsString, IsUUID } from 'class-validator';
+import { IsArray, IsIn, IsOptional, IsString, IsUUID } from 'class-validator';
 
 export class CreateAccessDto {
   @IsOptional()
@@ -16,10 +16,8 @@ export class CreateAccessDto {
   @IsUUID()
   granteeUserId?: string;
 
-  /**
-   * @deprecated Use the scopes instead
-   */
-  @IsEnum(AccessPermission, { each: true })
+  @IsArray()
+  @IsIn(Object.values(scopes), { each: true })
   @IsOptional()
-  permissions?: AccessPermission[];
+  scopes?: Scope[];
 }